// generated/src/app/api-models/bi-team-perf-management.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class BiTeamPerfManagement {
  issueId: number;
  title: string;
  status: string;
  dateReceived: Date;
  completionDate: Date;
  configurationRequestStatus: string;
  priority: string;
  assignedTo: string;
  category: string;
  requestor: string;
  requestingDepartment: string;
  complexity: string;
  application: string;
  workHours: number;
  itemType: number;
  path: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}